moved details to navigation pills
