+++ frontend/Site/src/app/features/details/dto/SiteDetailsMode.ts
@@ -1,0 +1,6 @@
+export enum SiteDetailsMode
+{
+   edit ="edit",
+   sr = "sr",
+   normal = "normal"
+}
