+++ frontend/Site/src/app/features/details/summary/Summary.css
@@ -1,0 +1,6 @@
+.summary-section-details
+{
+    gap: var(--layout-margin-xxlarge);
+    display: flex;
+    flex-direction: column;
+}
