+++ frontend/Site/src/app/features/site/dto/SiteState.ts
@@ -1,5 +1,6 @@
 import { ChangeTracker } from "../../../components/common/IChangeType";
 import { RequestStatus } from "../../../helpers/requests/status";
+import { SiteDetailsMode } from "../../details/dto/SiteDetailsMode";
 import { SiteResultDto, Sites } from "./Site"
 
 export class SiteState 
@@ -20,4 +21,5 @@ export class SiteState
     siteDetailsAddedStatus:string =  RequestStatus.idle;   
     siteDetailsUpdateStatus:string  = RequestStatus.idle;
     changeTracker: ChangeTracker[] = [];
+    siteDetailsMode: SiteDetailsMode = SiteDetailsMode.normal;
 }
